% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_data.R
\name{assessment_data}
\alias{assessment_data}
\title{Get raw assessment data for a single assessment_id}
\usage{
assessment_data(api, assessment_id)
}
\arguments{
\item{api}{An httr2 response object created with init_api()}

\item{assessment_id}{Integer. A single assessment_id}
}
\value{
A list of raw assessment data
}
\description{
Returns full JSON response from the Red List API as nested lists
}
\examples{
\dontrun{
assessment_data(api, assessment_id = 123)
}
}
