% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments_by_group.R
\name{assessments_by_group}
\alias{assessments_by_group}
\title{Get minimal assessment data for an API group}
\usage{
assessments_by_group(
  api,
  group,
  code,
  year_published = NULL,
  latest = TRUE,
  scope_code = NULL,
  wait_time = 0.5,
  show_warnings = TRUE
)
}
\arguments{
\item{api}{An httr2 response object created with init_api().}

\item{group}{String. A valid API group name. Valid names can be found by calling \code{list_group_names()}.}

\item{code}{String. A valid group code. Valid codes for an API group can be
found by calling e.g. \code{list_codes(api, "habitats")}}

\item{year_published}{Integer. The publication year you wish to filter.}

\item{latest}{Boolean. Return latest or historic assessments. Defaults to \code{TRUE}.}

\item{scope_code}{String. A valid assessment scope. Defaults to \code{NULL} (i.e. returns assessments of all scopes).
Valid scope codes can be found by calling \code{list_codes(api, "scopes")}.}

\item{wait_time}{Float. Time in seconds to wait between API calls. A wait time of >=0.5s will avoid API rate limiting.}

\item{show_warnings}{Boolean. Whether to show a warning to the console if no assessments are found for your specified arguments.
Defaults to \code{TRUE}.}
}
\value{
Returns a \code{tibble()} of minimal assessment data for your specified arguments. The minimal assessment data
provides assessment_ids and sis_taxon_ids which can be used with \code{assessment_data()}, \code{assessment_data_many()} and
\code{parse_assessment_data()} to get full assessment data.
}
\description{
The \code{assessment_by_} functions including \code{assessments_by_group()} return 'minimal' assessment data
for your specified filters (arguments). The minimal assessment data provides assessment_ids and sis_taxon_ids
(as a tibble) which can then be used with \code{assessment_data()}, \code{assessment_data_many()} and \code{partial_assessment_data()}
to get full assessment data.

The \code{assessments_by_group()} function acts upon the API assessment group names and codes.
Valid names can be found by calling \code{list_group_names()} and codes can be
found with \code{list_codes(api, "scopes")}.
}
\examples{
\dontrun{
assessments_by_group(api,
  group = "habitats",
  code = "1_1",
  year_published = 2012,
  latest = TRUE,
  scope_code = 1,
  wait_time = 0.5,
  show_warnings = TRUE)
}
}
