% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments_by_name.R
\name{assessments_by_name}
\alias{assessments_by_name}
\title{Get minimal assessment data for a Latin binomial}
\usage{
assessments_by_name(api, genus, species)
}
\arguments{
\item{api}{An httr2 response object created with init_api().}

\item{genus}{String. A genus name.}

\item{species}{String. A species name.}
}
\value{
Returns a \code{tibble()} of minimal assessment data for the supplied genus and species name. The minimal assessment data
provides assessment_ids and sis_taxon_ids which can then be used with \code{assessment_data()}, \code{assessment_data_many()} and
\code{partial_assessment_data()} to get full assessment data.
}
\description{
The \code{assessment_by_} functions including \code{assessments_by_name()} return 'minimal' assessment data
for your specified filters (arguments). The minimal assessment data provides assessment_ids and sis_taxon_ids
(as a tibble) which can be used with \code{assessment_data()}, \code{assessment_data_many()} and \code{partial_assessment_data()}
to get full assessment data.

The \code{assessments_by_name()} function returns minimal assessment data for a supplied Latin binomial.
The genus and species arguments are not case sensitive.
}
\examples{
\dontrun{
assessments_by_name(api, "panthera", "leo")
}
}
