% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments_by_sis_id.R
\name{assessments_by_sis_id}
\alias{assessments_by_sis_id}
\title{Get minimal assessment data for an SIS taxon ID}
\usage{
assessments_by_sis_id(api, sis_taxon_id)
}
\arguments{
\item{api}{An httr2 response object created with init_api().}

\item{sis_taxon_id}{Integer. A valid SIS taxon ID.}
}
\value{
Returns a \code{tibble()} of minimal assessment data for the supplied SIS taxon ID. The minimal assessment data
shows assessment_ids for the supplied SIS taxon ID which can then be used with \code{assessment_data()},
\code{assessment_data_many()} and \code{partial_assessment_data()} to get full assessment data.
}
\description{
The \code{assessment_by_} functions including \code{assessments_by_sis_id()} return 'minimal' assessment data
for your specified filters (arguments). The minimal assessment data provides assessment_ids and sis_taxon_ids
(as a tibble) which can be used with \code{assessment_data()}, \code{assessment_data_many()} and \code{partial_assessment_data()}
to get full assessment data.

The \code{assessments_by_sis_id()} function returns minimal assessment data for a supplied SIS taxon ID.
}
\examples{
\dontrun{
assessments_by_sis_id(api, 230)
}
}
