% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments_by_taxonomy.R
\name{assessments_by_taxonomy}
\alias{assessments_by_taxonomy}
\title{Get minimal assessment data for a taxonomic level}
\usage{
assessments_by_taxonomy(
  api,
  level,
  name,
  year_published = NULL,
  latest = TRUE,
  scope_code = NULL,
  wait_time = 0.5
)
}
\arguments{
\item{api}{An httr2 response object created with init_api().}

\item{level}{String. The level of taxonomic hierachy. Must be one of: kingdom, phylum, class, order or family.}

\item{name}{String. The name of taxonomic hierarchy.}

\item{year_published}{Integer. The publication year you wish to filter.}

\item{latest}{Boolean. Return latest or historic assessments. Defaults to \code{TRUE}.}

\item{scope_code}{String. A valid assessment scope. Defaults to \code{NULL} (i.e. returns assessments of all scopes).
Valid scope codes can be found by calling \code{list_codes(api, "scopes")}.}

\item{wait_time}{Time in seconds to wait between API calls. A wait of 0.5 seconds or greater will ensure you
won't hit the IUCN Red List API rate limit.
Defaults to \code{TRUE}.}
}
\value{
Returns a \code{tibble()} of minimal assessment data for your specified arguments. The minimal assessment data
provides assessment_ids and sis_taxon_ids which can be used with \code{assessment_data()}, \code{assessment_data_many()} and
\code{parse_assessment_data()} to get full assessment data.
}
\description{
The \code{assessment_by_} functions including \code{assessments_by_taxonomy()} return 'minimal' assessment data
for your specified filters (arguments). The minimal assessment data provides assessment_ids and sis_taxon_ids
(as a tibble) which can then be used with \code{assessment_data()}, \code{assessment_data_many()} and \code{parse_assessment_data()}
to get full assessment data.

The \code{assessments_by_taxonomy()} returns minimal assessment data for a given taxonomic hierarchy name.
}
\examples{
\dontrun{
assessments_by_taxonomy(api,
  level = "family",
  name = "felidae",
  year_published = 2022,
  latest = TRUE,
  scope_code = 1,
  wait_time = 0.5
)
}
}
