% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_names.R
\name{element_names}
\alias{element_names}
\title{Show element names}
\usage{
element_names()
}
\value{
A character vector of valid element names.
}
\description{
Returns a vector of element names which can be passed to \code{extract_element()}
to concatenate data from a large list of assessment data.
}
\examples{
\dontrun{
list_group_names()
}
}
