% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_element.R
\name{extract_element}
\alias{extract_element}
\title{Extract data from a named element}
\usage{
extract_element(assessments_list, element_name)
}
\arguments{
\item{assessments_list}{A list containing full assessment data.}

\item{element_name}{String.}
}
\value{
Returns a \code{tibble()} of
}
\description{
This is a helper function to return a tidy \code{tibble()} for a
specified element name. Here, an 'element name' is one of the named tibbles
created when running \code{parse_assessement_data()} or \code{assessment_data_many()}.

To get a list of valid element names that you can concatenate with \code{extract_element()}
first call \code{element_names()}. Pass one of these names to \code{extract_element()} to
create a tidy \code{tibble()} of that specific data.
}
\examples{
\dontrun{
extract_element(assessments_list, "habitats")
extract_element(assessments_list, "biogeographical_realms")
extract_element(assessments_list, "taxon_synonyms")
extract_element(assessments_list, "threats")
}
}
