% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_data_all.R
\name{gss_data_all}
\alias{gss_data_all}
\title{Fetch all green status of species (gss) assessment data}
\usage{
gss_data_all(api)
}
\arguments{
\item{api}{An \code{{httr2}} response object created with \code{init_api()}.}
}
\value{
Returns an unprocessed named list of assessment data for all
green status of species assessments.
The raw JSON from the gss endpoint has been coerced to a list but is otherwise
an exact representation of the response from the assessment endpoint.
}
\description{
Returns full assessment data for all green status of species assessments.
}
\examples{
\dontrun{
gss_data_all(api)
}
}
