% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-binary-infix.R
\name{\%||\%}
\alias{\%||\%}
\title{Binary infix operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value.}

\item{y}{A value.}
}
\value{
If x is NULL, will return y; otherwise returns x.
}
\description{
Binary infix operator
}
\keyword{internal}
