% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_api.R
\name{init_api}
\alias{init_api}
\title{Creates a request object for a valid red_list_api_key}
\usage{
init_api(red_list_api_key)
}
\arguments{
\item{red_list_api_key}{A v4 API token string (alphanumeric)}
}
\value{
An httr2 request object.
}
\description{
Creates a request object for a valid red_list_api_key
}
\examples{
api <- init_api("your_red_list_api_key")
}
