% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_red_list_api}
\alias{init_red_list_api}
\title{Creates a request object for a valid red_list_api_key}
\usage{
init_red_list_api(red_list_api_key)
}
\arguments{
\item{red_list_api_key}{A number.}
}
\value{
An httr2 request object.
}
\description{
Creates a request object for a valid red_list_api_key
}
\examples{
init_red_list_api('your_red_list_api_key')
}
