% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initialise_red_list_client}
\alias{initialise_red_list_client}
\title{Creates a request object for a valid red_list_api_key}
\usage{
initialise_red_list_client(red_list_api_key)
}
\arguments{
\item{red_list_api_key}{A string.}
}
\value{
A list.
}
\description{
Creates a request object for a valid red_list_api_key
}
