% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucnredlist-package.R
\docType{package}
\name{iucnredlist-package}
\alias{iucnredlist}
\alias{iucnredlist-package}
\title{iucnredlist: An R package for the IUCN Red List API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package for interacting with the IUCN Red List API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IUCN-UK/iucnredlist}
  \item \url{https://iucn-uk.github.io/iucnredlist/}
  \item Report bugs at \url{https://github.com/IUCN-UK/iucnredlist/issues}
}

}
\author{
\strong{Maintainer}: Simon Tarr \email{simon.tarr@iucn.org} (\href{https://orcid.org/0000-0001-8464-1240}{ORCID})

Authors:
\itemize{
  \item Ian Townsend \email{ian.townsend@iucn.org} (\href{https://orcid.org/0009-0003-8894-6019}{ORCID})
}

Other contributors:
\itemize{
  \item International Union for Conservation of Nature [copyright holder, funder]
}

}
\keyword{internal}
