% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_group_names.R
\name{list_group_names}
\alias{list_group_names}
\title{List API Endpoint Groups}
\usage{
list_group_names()
}
\value{
A character vector of API groups
}
\description{
Returns a list of API groups which have code/name parameters. These names can
be passed to \code{list_codes()} to get a list of codes or names for
an API group of interest (e.g. \code{list_codes(api, "threats")})
}
\examples{
\dontrun{
list_group_names()
}
}
