% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gss_data.R
\name{parse_gss_data}
\alias{parse_gss_data}
\title{Parse raw green status of species (gss) assessment data}
\usage{
parse_gss_data(raw_gss_data)
}
\arguments{
\item{raw_gss_data}{A \code{list()} generated by \code{gss_data_all()}.}
}
\value{
Returns a tibble of green status assessment data
}
\description{
Transforms a list of raw gss assessment data from \code{gss_data_all()} and
converts to a tibble with a single row per assessment.
The purpose of this function is to generate a 'tidy' tibble from the raw response,
which we feel is easier to work with in downstream analyses.

Due to the complexity of the taxon response, we have only included basic taxonomic
information to enable filtering on taxonomy or for species level.

The more extended text fields 'justification', 'assessor_names', 'reviewer_names'
'contributors', 'facilitators', 'compilers', 'url' and 'authority' have not been
included in the tibble returned from this function. If any of these are important
for your work, we would recommend fetching the data from the object created with
\code{gss_data_all()} and process according to your needs.
}
\examples{
\dontrun{
parse_gss_data(raw_gss_data)
}
}
