% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_version.R
\name{rl_version}
\alias{rl_version}
\title{Display API and Red List versions}
\usage{
rl_version(api)
}
\arguments{
\item{api}{A Red List API token string (alphanumeric)}
}
\value{
A list containing the API version and Red List version
}
\description{
Display API and Red List versions
}
\examples{
\dontrun{
api <- rl_version(api)
}
}
