% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_data.R
\name{assessment_data}
\alias{assessment_data}
\title{Fetch assessment data for a single assessment ID}
\usage{
assessment_data(api, assessment_id)
}
\arguments{
\item{api}{An \code{{httr2}} response object created with \code{init_api()}.}

\item{assessment_id}{Integer. A single Red List assessment_id.}
}
\value{
Returns an unprocessed named list of assessment data for a single assessment_id.
The raw JSON from the assessment endpoint has been coerced to a list but is
otherwise an exact representation of the response from the assessment endpoint.
}
\description{
Returns full assessment data for a single assessment_id as a named list.
Some of the elements within the top-level named list (e.g. taxon, conservation actions)
contains nested lists of data.
}
\examples{
\dontrun{
assessment_data(api, assessment_id = 266696959)
}
}
