% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_data.R
\name{assessment_data}
\alias{assessment_data}
\title{Get full assessment data for an assessment ID}
\usage{
assessment_data(api, assessment_id)
}
\arguments{
\item{api}{An httr2 response object created with init_api()}

\item{assessment_id}{Integer. A single assessment_id}
}
\value{
A list of length 1 containing full assessment data
}
\description{
foo bar
}
\examples{
assessment_data(api, assessment_id = 123)
}
